+++ src/Data/Monoid/Abelian.hs
@@ -26,9 +26,9 @@ instance Ord a => Monoid (FreeAbelianMonoid a) where
     mempty = FreeAbelianMonoid (Map.empty)
 
 type instance AlgebraType0 FreeAbelianMonoid a = Ord a
-type instance AlgebraType  FreeAbelianMonoid m = (Monoid m, AbelianSemigroup m)
+type instance AlgebraType  FreeAbelianMonoid m = (Ord m, Monoid m, AbelianSemigroup m)
 instance FreeAlgebra FreeAbelianMonoid where
     returnFree a = FreeAbelianMonoid (Map.singleton a 1)
     foldMapFree g (FreeAbelianMonoid as) = Map.foldMapWithKey (\a n -> stimes n $ g a) as 
-    proof0 = Proof Dict
     proof  = Proof Dict
+    forget = Proof Dict