+++ ChangeLog.md
@@ -6,3 +6,5 @@
 ## Unreleased changes
 
 - Monadicity example
+- Breaking change: changed proofs in FreeAlgebra and FreeAlgebra1 class; now
+  using `FreeAlgebra` and `FreeAlgebra1` classes let us deduce an adjunction.